Improve tracking of initial blockchain catchup
The chunked downloads get added to the head of the chain and change
the local_height value, so there is no need to adjust the
catchup_progress there.

In on_get_header, progress was only being updated when the chain was
connected, which means that it was fully caught up. Moved the logic to
increment on each recieved header.

The single headers work backward from the tail of the chain so its not
accurate to describe the height as being local_height +
catchup_progress. Renamed and reworked the function to return the
number of blocks remaining to download.